requirements.txt
msal
python-dotenv
streamlit
pandas
PyPDF2
openpyxl
O365
langchain-groq